requirements.txt
astroid==2.15.8
dill==0.3.7
isort==5.11.5
lazy-object-proxy==1.9.0
load-dotenv==0.1.0
mccabe==0.7.0
platformdirs==4.0.0
pylint==2.17.7
python-dotenv==0.21.1
tomli==2.0.1
tomlkit==0.12.5
typed-ast==1.5.5
typing_extensions==4.7.1
wrapt==1.16.0
